+++ lib/snov/domain_search.rb
@@ -4,7 +4,9 @@ require 'camel_snake_struct'
 module Snov
   class DomainSearch
     Response = Class.new(CamelSnakeStruct)
-    Response.example(MultiJson.load(File.read("#{__dir__}/fake_client/get_v2_domain-emails-with-info.json")))
+    Dir["#{__dir__}/fake_client/get_v2_domain-emails-with-info/*.json"].each do |example_file|
+      Response.example(MultiJson.load(File.read(example_file)))
+    end
     include Enumerable
 
     attr_reader :client, :domain, :type, :limit